:hammer: Fixed index conversion problem detected with clang
https://github.com/Gjacquenot/test_clang/runs/7747483505?check_suite_focus=true

[616/908] Building CXX object grpc/CMakeFiles/grpc.dir/src/GRPCForceModel.cpp.o
FAILED: grpc/CMakeFiles/grpc.dir/src/GRPCForceModel.cpp.o
/usr/bin/clang++ -DGTEST_HAS_POSIX_RE=0 -I. -I/opt/share/code/grpc/inc -I/opt/share/code/ssc -I/opt/share/code/exceptions/inc -I/opt/share/code/external_data_structures/inc -I/opt/share/code/core/inc -I/opt/share/code/environment_models/inc -I/opt/share/code/hdb_interpolators/inc -I/opt/share/code/mesh/inc -isystem /opt/eigen -isystem /opt/boost/include -isystem /opt/grpc/include -isystem grpc -isystem /opt/share/code/yaml-cpp/src -isystem /opt/share/code/yaml-cpp/include -isystem /opt/share/code/yaml-cpp/include/yaml-cpp -Werror -pedantic -pedantic-errors -Wextra -Wall -Wunused-function -Wunused-label -Wunused-parameter -Wunused-value -Wunused-variable -fno-common -Wformat=2 -Winit-self -Wpacked -Wpointer-arith -Wmissing-declarations -Wmissing-format-attribute -Wsign-compare -Wstrict-aliasing=2 -Wundef -Wconversion -Wno-misleading-indentation -Woverloaded-virtual -Weffc++ -Wwrite-strings -Wfatal-errors -Wno-deprecated -Wvariadic-macros -fPIC -std=gnu++11 -O3 -DNDEBUG -MD -MT grpc/CMakeFiles/grpc.dir/src/GRPCForceModel.cpp.o -MF grpc/CMakeFiles/grpc.dir/src/GRPCForceModel.cpp.o.d -o grpc/CMakeFiles/grpc.dir/src/GRPCForceModel.cpp.o -c /opt/share/code/grpc/src/GRPCForceModel.cpp
/opt/share/code/grpc/src/GRPCForceModel.cpp:64:39: fatal error: implicit conversion changes signedness: 'int' to 'std::vector::size_type' (aka 'unsigned long') [-Wsign-conversion]
            commands.reserve(response.commands_size());
                     ~~~~~~~ ~~~~~~~~~^~~~~~~~~~~~~~~
1 error generated.

+++ code/grpc/src/GRPCForceModel.cpp
@@ -61,7 +61,7 @@ class GRPCForceModel::Impl
             throw_if_invalid_status<Input,GRPCForceModel>(input, "set_parameters", status);
             needs_wave_outputs = response.needs_wave_outputs();
             max_history_length = response.max_history_length();
-            commands.reserve(response.commands_size());
+            commands.reserve(static_cast<size_t>(response.commands_size()));
             std::copy(response.commands().begin(), response.commands().end(), std::back_inserter(commands));
             force_frame.frame = response.frame();
             force_frame.angle.phi = response.phi();
@@ -180,7 +180,6 @@ GRPCForceModel::GRPCForceModel(const GRPCForceModel::Input& input, const std::st
 GRPCForceModel::GRPCForceModel(const TR1(shared_ptr)<Impl>& pimpl_, const std::string& body_name_, const EnvironmentAndFrames& env) :
         ForceModel(pimpl_->get_input().name, pimpl_->get_commands(), pimpl_->get_transformation_to_model_frame(), body_name_, env),
         pimpl(pimpl_)
-
 {
 }
